observer/backend/pyproject.toml
[project]
name = "observer-backend"
version = "0.1.0"
description = "Market Observer + Trade Recommender backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "websockets",
    "pydantic",
    "pyyaml",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "httpx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests/unit", "tests/integration"]
pythonpath = ["src"]
markers = [
    "unit: fast isolated unit tests",
    "network: tests that call external APIs (skip with -m 'not network')",
    "integration: project-level integration tests",
]
